hr-evaluation-overview

+++ src/main/java/com/hrms/careerpathmanagement/dto/EmployeePotentialPerformanceDTO.java
@@ -11,6 +11,6 @@ public class EmployeePotentialPerformanceDTO {
     Integer employeeId;
     String fullName;
     String profileImgUri;
-    Float potential;
+    Integer potential;
     Float performance;
 }
